requirements.txt
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
scipy==1.6.1
seaborn==0.11.1
six @ file:///tmp/build/80754af9/six_1605205327372/work
torch==1.7.1+cu110
torch-dct==0.1.5
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.57.0
typing-extensions==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
